@@ -1,6 +1,6 @@
 package com.resocoder.timertutorial
 
-import android.support.v7.app.AppCompatActivity
+import androidx.appcompat.app.AppCompatActivity
 import android.os.Bundle
 import kotlinx.android.synthetic.main.activity_timer.*
 
