@@ -1,7 +1,7 @@
 package com.resocoder.timertutorial
 
 import android.os.Bundle
-import android.support.v7.preference.PreferenceFragmentCompat
+import androidx.preference.PreferenceFragmentCompat
 
 
 class SettingsActivityFragment : PreferenceFragmentCompat() {
