@@ -11,7 +11,7 @@ import android.graphics.Color
 import android.media.RingtoneManager
 import android.net.Uri
 import android.os.Build
-import android.support.v4.app.NotificationCompat
+import androidx.core.app.NotificationCompat
 import com.resocoder.timertutorial.AppConstants
 import com.resocoder.timertutorial.R
 import com.resocoder.timertutorial.TimerActivity
